refactor: 패키지 구조를 형식에 맞게 리팩터링 & 경로 설정 수정
