--- src/main/java/org/sunday/projectpop/auth/JwtTokenProvider.java
@@ -1,95 +1,0 @@
-package org.sunday.projectpop.auth;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.JwtException;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Component;
-
-import javax.crypto.SecretKey;
-import java.time.Instant;
-import java.util.Date;
-import java.util.List;
-
-@Slf4j
-@Component
-public class JwtTokenProvider {
-
-    @Value("${jwt.secret}")
-    private String secretKey;
-
-    @Value("${jwt.expiration-ms}")
-    private long expirationMs;
-
-    public SecretKey getSecretKey() {
-        byte[] keyBytes = secretKey.getBytes();
-        log.debug("🛡️ JWT 시크릿 키 길이 (bytes): {}", keyBytes.length);
-        return Keys.hmacShaKeyFor(keyBytes);
-    }
-
-    public String generateToken(Authentication authentication, List<String> roles) {
-        String username = authentication.getName();
-        Instant now = Instant.now();
-        Date expiration = new Date(now.toEpochMilli() + expirationMs);
-
-        Claims claims = Jwts.claims()
-                .subject(username)
-                .add("roles", roles)
-                .build();
-
-        return Jwts.builder()
-                .subject(username)
-                .issuedAt(Date.from(now))
-                .expiration(expiration)
-                .claims(claims)
-                .signWith(getSecretKey(), Jwts.SIG.HS256)
-                .compact();
-    }
-
-    public String getUsername(String token) {
-        return Jwts.parser()
-                .verifyWith(getSecretKey())
-                .build()
-                .parseSignedClaims(token)
-                .getPayload()
-                .getSubject();
-    }
-
-    public boolean validateToken(String token) {
-        try {
-            Jwts.parser()
-                    .verifyWith(getSecretKey())
-                    .build()
-                    .parseSignedClaims(token);
-            return true;
-        } catch (JwtException | IllegalArgumentException e) {
-            return false;
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    public List<String> getRoles(String token) {
-        return (List<String>) Jwts.parser()
-                .verifyWith(getSecretKey())
-                .build()
-                .parseSignedClaims(token)
-                .getPayload()
-                .get("roles");
-    }
-
-    public Authentication getAuthentication(String token) {
-        UserDetails user = new User(getUsername(token), "",
-                getRoles(token).stream()
-                        .map(role -> new SimpleGrantedAuthority("ROLE_" + role))
-                        .toList());
-
-        return new UsernamePasswordAuthenticationToken(user, null, user.getAuthorities());
-    }
-}
