--- src/main/java/org/sunday/projectpop/dto/LoginDTO.java
@@ -1,6 +1,0 @@
-package org.sunday.projectpop.dto;
-
-public record LoginDTO(
-        String email,
-        String password
-) {}
