--- src/main/java/org/sunday/projectpop/dto/ProfileResponse.java
@@ -1,6 +1,0 @@
-package org.sunday.projectpop.dto;
-
-public record ProfileResponse(
-        UserProfileResponseDTO profile,
-        UserTagCacheDTO tags
-) {}
