--- src/main/java/org/sunday/projectpop/dto/UserProfileResponseDTO.java
@@ -1,13 +1,0 @@
-package org.sunday.projectpop.dto;
-
-import java.util.UUID;
-
-// 📌 조회용 (userId 포함)
-public record UserProfileResponseDTO(
-        UUID userId,
-        String email,
-        String nickname,
-        String bio,
-        String profileImageUrl,
-        String phone
-) {}
