+++ src/main/java/org/sunday/projectpop/dto/auth/LoginDTO.java
@@ -1,0 +1,6 @@
+package org.sunday.projectpop.dto.auth;
+
+public record LoginDTO(
+        String email,
+        String password
+) {}
