+++ src/main/java/org/sunday/projectpop/dto/profile/ProfileResponse.java
@@ -1,0 +1,8 @@
+package org.sunday.projectpop.dto.profile;
+
+import org.sunday.projectpop.dto.tags.UserTagCacheDTO;
+
+public record ProfileResponse(
+        UserProfileResponseDTO profile,
+        UserTagCacheDTO tags
+) {}
