+++ src/main/java/org/sunday/projectpop/dto/profile/UserProfileResponseDTO.java
@@ -1,0 +1,13 @@
+package org.sunday.projectpop.dto.profile;
+
+import java.util.UUID;
+
+// 📌 조회용 (userId 포함)
+public record UserProfileResponseDTO(
+        UUID userId,
+        String email,
+        String nickname,
+        String bio,
+        String profileImageUrl,
+        String phone
+) {}
