+++ src/main/java/org/sunday/projectpop/entity/account/Users.java
@@ -1,0 +1,35 @@
+package org.sunday.projectpop.entity.account;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.UUID;
+
+@Table(name = "user_account")
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+public class Users {
+
+    @Id
+    @GeneratedValue
+    private UUID id;
+
+    @Column(nullable = false, unique = true, length = 100)
+    private String email;
+
+    @Column(nullable = false, length = 60)
+    private String password; // BCrypt 인코딩된 비밀번호
+
+    @Column(nullable = false, length = 20)
+    private String role = "USER";
+
+    @Column(nullable = false, length = 20)
+    private String provider = "local"; // local, kakao, github, google
+
+    @Column(length = 100)
+    private String providerId; // 소셜 로그인용 ID
+}
