--- src/main/java/org/sunday/projectpop/repository/UsersRepository.java
@@ -1,12 +1,0 @@
-package org.sunday.projectpop.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.sunday.projectpop.entity.Users;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface UsersRepository extends JpaRepository<Users, UUID> {
-    Optional<Users> findByEmail(String email);
-    boolean existsByEmail(String email);
-}
