+++ src/main/java/org/sunday/projectpop/repository/profile/UserProfileRepository.java
@@ -1,0 +1,9 @@
+package org.sunday.projectpop.repository.profile;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.sunday.projectpop.entity.profile.UserProfile;
+
+import java.util.UUID;
+
+public interface UserProfileRepository extends JpaRepository<UserProfile, UUID> {
+}
